[#285] fix : 프로필 올렸을때 수정 안되게 변경
