+++ src/recoilStores/type/MyPage/MyProfileType.ts
@@ -15,6 +15,7 @@ export interface MyProfileType {
     idealAge: string;
     idealType: string;
     level: string;
+    profileStatus: string; //ACTIVE, INACTIVE
 }
 
 export interface MyProfileResponseType {
